// frontend/src/app/features/business-partners/carriers/models/carrier-update-request.model.ts
export interface CarrierUpdateRequest {
  code: string;
  companyName: string;

  contactName: string | null;
  email: string | null;
  phoneNumber: string | null;

  address: string;
  city: string;
  postalCode: string | null;
  country: string;

  vatNumber: string | null;
  licenseNumber: string;

  active: boolean;
}